requirements.txt
ase==3.20.1
beautifulsoup4==4.11.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dgl-cpu==0.4.3.post2
dgl-gpu==0.4.3.post2
filelock==3.8.0
future==0.18.2
gdown==4.5.1
googledrivedownloader==0.4
h5py==2.10.0
idna==2.10
isodate==0.6.0
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.2.0
littleutils==0.2.2
llvmlite==0.32.1
MarkupSafe==1.1.1
matplotlib==3.3.2
networkx==2.5
numba==0.49.1
numpy==1.19.2
ogb==1.3.1
outdated==0.2.1
pandas==1.1.3
patsy==0.5.1
Pillow==7.2.0
Pillow-SIMD==7.0.0.post3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-louvain==0.13
pytz==2020.4
rdflib==5.0.0
requests==2.25.1
scikit-learn==0.23.0
scipy==1.4.1
six==1.15.0
soupsieve==2.3.2.post1
statsmodels==0.11.1
threadpoolctl==2.1.0
torch==1.5.0
torch-cluster==1.5.7
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.6
torch-spline-conv==1.2.0
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2